2021/10/27 last commit I'm Done!!!!

+++ app/src/main/java/com/example/risingproject/src/allReview/models/ReviewsItem.kt
@@ -7,7 +7,7 @@ data class ReviewsItem(
     @SerializedName("helpedNum") val helpedNum: Int,
     @SerializedName("itemId") val itemId: Int,
     @SerializedName("photo") val photo: String,
-    @SerializedName("reviewAvg") val reviewAvg: Int,
+    @SerializedName("reviewAvg") val reviewAvg: Float,
     @SerializedName("reviewId") val reviewId: Int,
     @SerializedName("uploadTime") val uploadTime: String,
     @SerializedName("userNickName") val userNickName: String
